Factor export insertion out of action
