+++ client/src/modules/blockEditor/actions.ts
@@ -1,4 +1,3 @@
-import * as uuidv1 from "uuid/v1";
 import { UPDATE_BLOCK } from "../blocks/actions";
 import * as slateChangeMutations from "../../slate-helpers/slate-change-mutations";
 
@@ -57,25 +56,6 @@ export const changePointerReference = ({ id, reference }) => {
   };
 };
 
-const onlyOneNodeThatIsPointerExport = nodes => {
-  return (
-    nodes.length === 1
-    &&
-    nodes[0].type === "pointerExport"
-  );
-};
-
-const twoNodesFirstExportSecondEmptyString = nodes => {
-  return (
-    nodes.length === 2
-    &&
-    nodes[0].type === "pointerExport"
-    &&
-    nodes[1].object === "text"
-    && nodes[1].leaves[0].text === ""
-  );
-};
-
 export const exportSelection = () => {
   return async (dispatch, getState) => {
     const { blocks, blockEditor } = await getState();
@@ -83,51 +63,11 @@ export const exportSelection = () => {
 
     const block = blocks.blocks.find(b => b.id === hoveredItem.blockId);
 
-    const uuid = uuidv1();
     if (block) {
-
-      const value = block.value;
-
-      const initialFragment = value.fragment;
-      const initialTopLevelNodes = initialFragment.nodes.toJSON()[0].nodes;
-
-      const isNestedInPointerExport =
-        onlyOneNodeThatIsPointerExport(initialTopLevelNodes)
-        ||
-        twoNodesFirstExportSecondEmptyString(initialTopLevelNodes);
-
-      const change = value.change();
-
-      // if we are exporting a nested pointer, guarantee that the selection is
-      // not extending to an edge, because selections on the edge of an inline
-      // pick up what comes after
-      if (isNestedInPointerExport) {
-        slateChangeMutations.moveSelectionAwayFromPointerEdge(change);
-      }
-
-      // A Slate fragment is a document: value.fragment is the document
-      // encompassing the currently selected portion of the editor.
-      // If the user has selected part of a deeply nested node, then the
-      // associated fragment will include all levels of the nesting.
-      // The next few lines (until the end of the while loop) drill down
-      // through such a nested document to unnest all the nodes that are
-      // selected.
-      const maybeChangedFragment = change.value.fragment;
-      const maybeChangedTopLevelNodes = maybeChangedFragment.nodes.toJSON()[0].nodes;
-
-      let nodes = maybeChangedTopLevelNodes;
-      while (onlyOneNodeThatIsPointerExport(nodes)) {
-        nodes = nodes[0].nodes;
-      }
-
-      change.insertInline({
-        type: "pointerExport",
-        data: { pointerId: uuid },
-        nodes,
-      });
-
+      const change = block.value.change();
+      slateChangeMutations.insertPointerExport(change);
       slateChangeMutations.normalizeExportSpacing(change);
-
+      
       dispatch({
         type: UPDATE_BLOCK,
         id: block.id,
