+++ client/src/slate-helpers/slate-change-mutations/index.js
@@ -1,5 +1,6 @@
 import { adjustCursorIfAtEdge } from "./adjustCursorIfAtEdge";
 import { handleCursorNavigationAcrossPointerEdge } from "./handleCursorNavigationAcrossPointerEdge";
+import { insertPointerExport } from "./insertPointerExport";
 import { moveSelectionAwayFromPointerEdge } from "./moveSelectionAwayFromPointerEdge";
 import { normalizeAfterRemoval } from "./normalizeAfterRemoval";
 import { normalizeExportSpacing } from "./normalizeChange";
@@ -7,6 +8,7 @@ import { normalizeExportSpacing } from "./normalizeChange";
 export {
   adjustCursorIfAtEdge,
   handleCursorNavigationAcrossPointerEdge,
+  insertPointerExport,
   moveSelectionAwayFromPointerEdge,
   normalizeAfterRemoval,
   normalizeExportSpacing
